% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.r
\name{dGammaSD}
\alias{dGammaSD}
\title{Compute gamma stage duration densities.}
\usage{
dGammaSD(mu, cv, nGrid, tMax)
}
\arguments{
\item{mu}{A vector of length equal to one minus the number of stages giving the mean stage durations
for all but the last stage.}

\item{cv}{A vector of the same length as \code{mu} giving the coefficient of variation, which is equal to one
over the square root of the shape, for all but the last stage.}

\item{nGrid}{The number of grid points to use.}

\item{tMax}{The maximum time point in the grid (i.e. the right endpoint).}
}
\value{
A matrix of stage duration probabilities where the first column gives the grid
        and the remaining columns give the stage duration probabilities.
}
\description{
The function is intended to be used in the calculation of numeric likelihoods.
}
\examples{
## Simulate gamma stage durations for ten individuals.
dGammaSD(mu = c(5,2,4), cv = c(1,.5,.3), nGrid = 20, tMax = 15)
}

