% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.r
\name{dWeibSD}
\alias{dWeibSD}
\title{Compute Weibull stage duration densities.}
\usage{
dWeibSD(shape, scale, nGrid, tMax)
}
\arguments{
\item{shape}{A vector of length equal to one minus the number of stages giving the shape of the stage durations
for all but the last stage.}

\item{scale}{A vector of length equal to one minus the number of stages giving the scale of the stage durations
for all but the last stage.}

\item{nGrid}{The number of grid points to use.}

\item{tMax}{The maximum time point in the grid (i.e. the right endpoint).}
}
\value{
A matrix of stage duration probabilities where the first column gives the grid
        and the remaining columns give the stage duration probabilities.
}
\description{
The function is intended to be used in the calculation of numeric likelihoods.
}
\examples{
## Compute Weibull stage duration probabilities over a grid of 20 points.
dWeibSD(shape = c(3,1,2), scale = c(3,10,5), nGrid = 20, tMax = 15)
}

