% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.r
\name{mnLogLik}
\alias{mnLogLik}
\title{Approximate Monte Carlo multinomial log likelihood for stage frequency data with known initial numbers.}
\usage{
mnLogLik(data, rStageDur, mortRate, ...)
}
\arguments{
\item{data}{A data matrix of class sfMat.}

\item{rStageDur}{A function for simulating stage durations. Should return a matrix where the columns
give the simulated durations for each stage (in order), and the rows represent a simulated individual.
Simulated durations may be equalt to Inf if an indiviudal does not mature from a stage.}

\item{mortRate}{A vector of mortality rates for each stage. Should be of the same length as
the number of columns returned by rStageDur.}

\item{...}{Further arguments passed to rStageDur.}
}
\value{
The approximate (stochastic) log-likelihood value.
}
\description{
Approximate Monte Carlo log likelihood for multinomial sampling with
known initial number of individuals at start of development.
}

