% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.r
\name{mnLogLik2}
\alias{mnLogLik2}
\title{Approximate multinomial log likelihood for stage frequency data with fixed number of
       individuals.}
\usage{
mnLogLik2(data, rStageDur, mortRate = NULL, ...)
}
\arguments{
\item{data}{A data matrix of class sfMat.}

\item{rStageDur}{A function for simulating stage durations. Should return a matrix where the columns
give the simulated durations for each stage (in order), and the rows represent a simulated individual.
Simulated durations may be equalt to Inf if an indiviudal does not mature from a stage.}

\item{mortRate}{A vector of mortality rates for each stage. If NULL (the default), mortality rates are
assumed to be zero.}

\item{...}{Further arguments passed to rStageDur.}
}
\value{
The approximate (stochastic) log-likelihood value.
}
\description{
Approximate multinomial log likelihood for stage frequency data with fixed number of
       individuals.
}

