% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.r
\name{mnLogLikN}
\alias{mnLogLikN}
\title{Approximate numerical multinomial log likelihood for stage frequency data with known initial numbers.}
\usage{
mnLogLikN(data, dStageDur, mortRate, ...)
}
\arguments{
\item{data}{A data matrix of class sfMat.}

\item{dStageDur}{A function computing stage duration densities. Should return a matrix where the firs column
give the grid over which numeric integration is to be performed and the remaining columns
give the density of the stage durations evaluated at the grid point.
Simulated durations may be equalt to Inf if an indiviudal does not mature from a stage.}

\item{mortRate}{A vector of mortality rates for each stage. Should be of the same length as
the number of columns returned by rStageDur.}

\item{...}{Further arguments passed to rStageDur.}
}
\value{
The approximate numeric log-likelihood value.
}
\description{
Approximate numerical log likelihood for multinomial sampling with
known initial number of individuals at start of development.
}

