% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.r
\name{rGammaSD}
\alias{rGammaSD}
\title{Simulate gamma distributed stage durations.}
\usage{
rGammaSD(mu, cv, N)
}
\arguments{
\item{mu}{A vector of length equal to one minus the number of stages giving the mean stage durations
for all but the last stage.}

\item{cv}{A vector of the same length as \code{mu} giving the coefficient of variation, which is equal to one
over the square root of the shape, for all but the last stage.}

\item{N}{The number of stage durations (individuals) to simulate.}
}
\value{
A matrix of stage durations where each row gives tha stage durations for one individual.
}
\description{
The function is intended to be used in the calculation of Monte Carlo likelihoods.
}
\examples{
## Simulate gamma stage durations for ten individuals.
rGammaSD(mu = c(5,2,4), cv = c(1,.5,.3), N = 10)
}

