% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.r
\name{rWeibSD}
\alias{rWeibSD}
\title{Simulate Weibull distributed stage durations.}
\usage{
rWeibSD(shape, scale, N)
}
\arguments{
\item{shape}{A vector of length equal to one minus the number of stages giving the shape of the stage durations
for all but the last stage.}

\item{scale}{A vector of length equal to one minus the number of stages giving the scale of the stage durations
for all but the last stage.}

\item{N}{The number of stage durations (individuals) to simulate.}
}
\value{
A matrix of stage durations where each row gives tha stage durations for one individual.
}
\description{
The function is intended to be used in the calculation of Monte Carlo likelihoods.
}
\examples{
## Simulate Weibull stage durations for ten individuals.
rWeibSD(shape = c(3,1,2), scale = c(3,10,5), N = 10)
}

