% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.r
\name{simMNCohort}
\alias{simMNCohort}
\title{Simulate stage frequency data with known initial number of individuals for each sampling occasion.}
\usage{
simMNCohort(sampleTimes, N0, rStageDur, mortRate, Narg, ...)
}
\arguments{
\item{sampleTimes}{Vector of times at which the cohort is sampled. Development is assumed to start at time 0.}

\item{N0}{An integer vector of length equal to the length of \code{sampleTimes}, or of length one, giving the known
number of individuals at the start of development. If the length is one, the value is taken as the number
initial  individuals for eac sampling time.}

\item{rStageDur}{A function simulating stage durations, and returning stage durations as a matrix where rows
represent individuals and columns represent stage durations for each stage. The function should
have an argument for the number of individuals to simulate.}

\item{mortRate}{A vector of mortality rates for each stage.}

\item{Narg}{The name of the argument to the function \code{rStageDur} which determines the number of individuals
that are simulated.}

\item{...}{Further arguments passed to \code{rStageDur}.}
}
\value{
An sfData object.
}
\description{
Simulate stage frequency data with known initial number of individuals for each sampling occasion.
}
\examples{
data = simMNCohort(seq(1,50, length.out=15), N0 = 30, rGammaSD, mu = c(5, 15, 8), cv = c(.5,.5,.5),
                   mortRate = c(0.01, 0.01,0.01,0.01), Narg = "N")
}

