\name{stageFreq-package}
\alias{stageFreq-package}
\alias{stageFreq}
\docType{package}
\title{
\packageTitle{stageFreq}
}
\description{
\packageDescription{stageFreq}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stageFreq}
\packageIndices{stageFreq}
}
\author{
\packageAuthor{stageFreq}

Maintainer: \packageMaintainer{stageFreq}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
