% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.r
\name{stageFreqMat}
\alias{stageFreqMat}
\title{Convert a data frame or matrix into an sfMat object.}
\usage{
stageFreqMat(data, timeVar, stageVars, N0 = NULL, gamma = NULL)
}
\arguments{
\item{data}{A matrix or data frame containing stage frequency data where each row gives the
sample at a specific time point.}

\item{timeVar}{The name of the column of \code{data} holding the time variable,
which defines at which time a sample is taken.}

\item{stageVars}{A vector of names of the columns of \code{data} holding the stage counts.
The vector should be ordered from the first to the last stage.}

\item{N0}{The name of the column giving the known initial number of individuals at each
sampling time. May be set to NULL if not known.}

\item{gamma}{The name of the column giving the sampling effort at each time.
Used by \code{\link{poisLogLik}}. May be set to NULL (the default).}
}
\value{
A data matrix of class sfMat.
}
\description{
Convert a data frame or matrix into an sfMat object, to be used with the likelihood functions
\code{\link{mnLogLik}}, \code{\link{mnLogLik2}} and \code{\link{poisLogLik}}.
}

