% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stageFreqMC}
\alias{stageFreqMC}
\title{Compute stage frequencies from simulated stage durations.}
\usage{
stageFreqMC(sampleTimes, sDur, mortRate)
}
\arguments{
\item{sampleTimes}{Vector of times at which the cohort was sampled.}

\item{sDur}{A matrix of stage durations where each row corresponds to the stage duration of an individual.
If there is no exit from the last stage, the stage duration of that stage can be set to Inf.}

\item{mortRate}{A vector of mortality rates for each stage. Should be of length equal to the number of
columns of sDur.}
}
\value{
A matrix of stage counts. Each row gives the number of individuals in each stage for a given sampling time.
}
\description{
Compute stage frequencies from simulated stage durations.
}

